config: temp disable ruleset to resolve errs
The configured GitHub repository rules does not allow code changes
written by GitHub Actions to be pushed due to unsigned commits and
required status checks.

An alternative will be set up a custom bot user using GitHub App so the
author can be whitelisted but the changes might be too excessive. In the
meantime, we will temporarily allow disable the enforcement of those
rules.